Fixed makefile for Yocto
mxc_v4l2_capture.ko should not be compiled in yocto, as it will conflict
with the one built as part of the kernel tree. Instead it should be applied
as a patch.

mxc_v4l2_capture.ko can be built independantly by specifying
OUT_OF_TREE_BUILD=y in the environment before building.

KERNEL_BUILD was switched to KBUILD_OUTPUT for compatibility with Yocto.
They are equivalent.